moved contracted week to vacation utils; called in to handler now since splitVacationDyasByYear used in more than one place.
